requirements.txt
openai==0.28.1
gitpython==3.1.40
flask==3.0.0
python-dotenv==1.0.0
schedule==1.2.0

